Add files via upload
This file includes three section. The name of the first section is Run_CIFAR_10 and it consists of the all codes to implement the ResNet to predict the rotation by multiplies of 90 and 45 degrees on CIFAR-10 dateset. Similarly, the second and third sections consist of the  all codes to implement the ResNet on CIFAR-100 and Tinyimagenet-200 datasets, respectively.